feat: Implement elitism and allow odd pop size
Before, we calculated the number of genetic modification procedures
to be the length of old population divided by 2.
This isn't good when the population size is odd, since using int()
on that number would always give us the floor.
Consider the example with 5 snakes in a population, 5/2=2.5,
int(2.5) = 2. When we have 2 genetic procedures, it would create
4 new snakes, which would mean that our old pop and new pop sizes
aren't the same.