fix: Update population_size, fix old body draw
Two things were fixed with this commit.
The first one is that we didn't update population_size variable
inside of the new_population, which led to occasional
index out-of-bounds error.
The second is that we have always drawn the body of a previous
generation's chosen snake at the start of the new generation.